requirements.txt
# Web Application
Flask
flask-cors
flask-apispec==0.8.0
bjoern

# Schemas
marshmallow
marshmallow-sqlalchemy
flask-marshmallow
simplejson

# Database
psycopg2
Flask-SQLAlchemy
Flask-Migrate
flask-redis
sqlalchemy-utils

# Utils
Flask-Script
flask-bcrypt

# Quality
pytest-flask
pytest==5.00
pytest-cov
flake8
pycodestyle
coveralls